refactor makeTaskSpec
