+++ master/internal/command/api.go
@@ -8,8 +8,6 @@ import (
 	"github.com/determined-ai/determined/master/internal/db"
 	"github.com/determined-ai/determined/master/pkg/actor"
 	"github.com/determined-ai/determined/master/pkg/actor/api"
-	"github.com/determined-ai/determined/master/pkg/model"
-	"github.com/determined-ai/determined/master/pkg/tasks"
 )
 
 // RegisterAPIHandler initializes and registers the API handlers for all command related features.
@@ -19,37 +17,27 @@ func RegisterAPIHandler(
 	db *db.PgDB,
 	proxyRef *actor.Ref,
 	timeout int,
-	defaultAgentUserGroup model.AgentUserGroup,
-	makeTaskSpec tasks.MakeTaskSpecFn,
 	middleware ...echo.MiddlewareFunc,
 ) {
 	system.ActorOf(actor.Addr("commands"), &commandManager{
-		defaultAgentUserGroup: defaultAgentUserGroup,
-		db:                    db,
-		makeTaskSpec:          makeTaskSpec,
+		db: db,
 	})
 	echo.Any("/commands*", api.Route(system, nil), middleware...)
 
 	system.ActorOf(actor.Addr("notebooks"), &notebookManager{
-		defaultAgentUserGroup: defaultAgentUserGroup,
-		db:                    db,
-		makeTaskSpec:          makeTaskSpec,
+		db: db,
 	})
 	echo.Any("/notebooks*", api.Route(system, nil), middleware...)
 
 	system.ActorOf(actor.Addr("shells"), &shellManager{
-		defaultAgentUserGroup: defaultAgentUserGroup,
-		db:                    db,
-		makeTaskSpec:          makeTaskSpec,
+		db: db,
 	})
 	echo.Any("/shells*", api.Route(system, nil), middleware...)
 
 	system.ActorOf(actor.Addr("tensorboard"), &tensorboardManager{
-		defaultAgentUserGroup: defaultAgentUserGroup,
-		db:                    db,
-		makeTaskSpec:          makeTaskSpec,
-		proxyRef:              proxyRef,
-		timeout:               time.Duration(timeout) * time.Second,
+		db:       db,
+		proxyRef: proxyRef,
+		timeout:  time.Duration(timeout) * time.Second,
 	})
 	echo.Any("/tensorboard*", api.Route(system, nil), middleware...)
 }
